+++ src/Controller/Cloud/API/FileController.php
@@ -23,21 +23,17 @@ class FileController
      */
     public function GET_file(FileManager $fm, Notifications $notifications)
     {
+        // Get params
         $request = Request::createFromGlobals();
+        $data = $request->query->all(); // url passed params
+        if(isset($data['url'])) $path = $data['url']; else return $notifications->JSONResponse(400, false, 'You must provide relative file path.');
 
-        // Retrieve params
-        if (!empty(($request->request->get('url')))) { 
-            $path = $request->request->get('url');  // body passed params
-        } else if (!empty($request->query->get('url'))) {
-            $path = $request->query->get('url');    // url passed params
-        } else {
-            return $notifications->JSONResponse(400, false, 'You must provide relative file path.');
-        }
-
+        // Check file exists
         if(!$fm->file_exists($path)) {
-            return $notifications->JSONResponse(404, '<code>' . $path . '</code> was not found in the server.');
+            return $notifications->JSONResponse(404, false, '<code>' . $path . '</code> was not found in the server.');
         }
 
+        // Return file as binary
         $response = new BinaryFileResponse($fm->file_path($path));
 
         $disposition = HeaderUtils::makeDisposition(
@@ -59,26 +55,23 @@ class FileController
      */
     public function POST_file(FileManager $fm, Notifications $notifications)
     {
+        // Get params
         $request = Request::createFromGlobals();
+        $data = $request->request->all(); // form passed params
+        if(isset($data['url'])) $path = $data['url']; else return $notifications->JSONResponse(400, false, 'You must provide relative file path.');
+        $files = $request->files->all();
 
-        // Retrieve params
-        if (!empty(($request->request->get('url')))) { 
-            $path = $request->request->get('url');  // body passed params
-        } else if (!empty($request->query->get('url'))) {
-            $path = $request->query->get('url');    // url passed params
-        } else {
-            return $notifications->JSONResponse(400, false, 'You must provide relative file path.');
-        }
-        $file = $request->files->get('file');
-
-        if(!empty($file)) {
+        if(!empty($files)) {
             // move uploaded file
-            $fm->move_uploded_file($file, $path);
-            return $notifications->JSONResponse(202, '<code>' . $file . '</code> was successfully uploaded in the server.');
+            foreach ($files as $file){
+                var_dump($file);
+                $fm->move_uploded_file($file, $path);
+            }
+            return $notifications->JSONResponse(202, false, '<code>[' . implode(",", $files) . ']</code> were successfully uploaded in the server.');
         } else {
             // create empty file
             $fm->create_file($path);
-            return $notifications->JSONResponse(201, '<code>' . $path . '</code> was successfully created in the server.');
+            return $notifications->JSONResponse(201, false, '<code>' . $path . '</code> was successfully created in the server.');
         }
     }
 
@@ -91,44 +84,32 @@ class FileController
      * @param content <STRING> (Optional) The new file content.
      */
     public function PUT_file(FileManager $fm, Notifications $notifications) {
+        // Get params
         $request = Request::createFromGlobals();
-
-        // Retrieve path param
-        if (!empty(($request->request->get('url')))) { 
-            $path = $request->request->get('url');  // body passed params
-        } else if (!empty($request->query->get('url'))) {
-            $path = $request->query->get('url');    // url passed params
-        } else {
-            return $notifications->JSONResponse(400, false, 'You must provide relative file path.');
-        }
-
-        // Retrieve newpath param
-        if (!empty(($request->request->get('newurl')))) { 
-            $newpath = $request->request->get('newurl');  // body passed params
-        } else if (!empty($request->query->get('newurl'))) {
-            $newpath = $request->query->get('newurl');    // url passed params
-        }
+        $data = json_decode($request->getContent(), true); // json body passed params
+        if(isset($data['url'])) $path = $data['url']; else return $notifications->JSONResponse(400, false, 'You must provide relative file path.');
+        if(isset($data['newurl'])) $newpath = $data['newurl'];
+        if(isset($data['content'])) $content = $data['content'];
 
         // Rename file
-        if(!empty($newpath)) {
-            if($fm->rename($path, $newpath))
-                return $notifications->JSONResponse(202, '<code>' . $path . '</code> was successfully renamed as <code>' . $newpath . '</code>.');
-        }
-
-        // Retrieve content param
-        if (!empty(($request->request->get('content')))) { 
-            $content = $request->request->get('content');  // body passed params
-        } else if (!empty($request->query->get('content'))) {
-            $content = $request->query->get('content');    // url passed params
+        if(isset($data['newurl'])) {
+            if($fm->rename($path, $newpath)) {
+                return $notifications->JSONResponse(202, false, '<code>' . $path . '</code> was successfully renamed as <code>' . $newpath . '</code>.');
+            } else {
+                return $notifications->JSONResponse(500, false, 'An error occured will trying to rename <code>' . $path . '</code> into <code>' . $newpath . '</code>.');
+            }
         }
 
         // Update file content
-        if(!empty($content)) {
-            if($fm->update_file($path, $content))
-                return $notifications->JSONResponse(202, 'Content of <code>' . $path . '</code> was successfully updated.');
+        if(isset($data['content'])) {
+            if($fm->update_file($path, $content)) {
+                return $notifications->JSONResponse(202, false, 'Content of <code>' . $path . '</code> was successfully updated.');
+            } else {
+                return $notifications->JSONResponse(500, false, 'An error occured will trying to update file content.');
+            }
         }
 
-        return $notifications->JSONResponse(500, false, 'An error occured will trying to update file.');
+        return $notifications->JSONResponse(400, false, 'Please provide new file path (<code>newurl</code>) to move/rename it or new file content (<code>content</code>) to update it.');
     }
 
     /**
@@ -138,21 +119,16 @@ class FileController
      * @param url <PATH> The new file relative url to be created in the server.
      */
     public function DELETE_file(FileManager $fm, Notifications $notifications) {
+        // Get params
         $request = Request::createFromGlobals();
-
-        // Retrieve path param
-        if (!empty(($request->request->get('url')))) { 
-            $path = $request->request->get('url');  // body passed params
-        } else if (!empty($request->query->get('url'))) {
-            $path = $request->query->get('url');    // url passed params
-        } else {
-            return $notifications->JSONResponse(400, false, 'You must provide relative file path.');
-        }
+        $data = json_decode($request->getContent(), true); // json body passed params
+        if(isset($data['url'])) $path = $data['url']; else return $notifications->JSONResponse(400, false, 'You must provide relative file path.');
 
         // Delete file
-        if($fm->delete_file($path))
+        if($fm->delete_file($path)) {
             return $notifications->JSONResponse(200, false, '<code>' . $path . '</code> was successfully deleted.');
-        else
-            return $notifications->JSONResponse(500, false, 'An error occured will trying to delete file.');
+        }
+
+        return $notifications->JSONResponse(500, false, 'An error occured will trying to delete file.');
     }
 }
